backend/pyproject.toml
[project]
name = "llm-wiki-backend"
version = "0.1.0"
description = "Backend API for the LLM Wiki professional MVP"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pgvector>=0.3.6",
  "sqlalchemy>=2.0.30",
  "psycopg[binary]>=3.2.0",
  "pydantic-settings>=2.4.0",
  "python-multipart>=0.0.9",
  "pypdf>=5.0.0",
  "python-docx>=1.1.0",
  "numpy>=1.26.0",
  "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.3.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
